+++ src/web/assets/uploader/webpack.config.js
@@ -1,0 +1,11 @@
+const { getConfig } = require("@craftcms/webpack");
+const CopyWebpackPlugin = require("copy-webpack-plugin");
+
+module.exports = getConfig({
+  context: __dirname,
+  config: {
+    entry: {
+      Uploader: "./Uploader.js",
+    },
+  },
+});